refactor(poolcardtri/managestake): separate Stake managing into separate component
